refactor: bus를 shuttle로 변경하여 불분명하거나 겹치는 네이밍 수정
